User story #2 finding the hallways
To find hallways we would have to find all the empty spaces around the current pick the element with lower index and commit it to the cache and move forward until we hit a dead end. Thats achieved in 2 steps
a. Start from the entrance and mark the node in memory.
b. Move to all the adjacent nodes and continue this loop until there is no unmarked adjacent node while memorizing the path.