User story #3 finding the exit of a room
Multiple hallways stacked together becomes a room or to put it simply multiple hallways with dead ends form a room. To navigate out of it we need to figure where the dead ends are and what alternate routes can be found. To achieve this we need to:
a. If we've already been there or there is a wall, quit
b. Fill up the path travelled with a wall and check for a new dead ends