User story #4 and #5 and error handling
Compounded rooms (snaking hallways or rooms) will become a maze or and exit of one rooms will be entrance of the other. The DFS method of searching till the end for a dead end and looking for alternatives will pass through one room after another till reaches the exit.

And some error handling so that the dfs will not cross the maze's boundaries.